refactor(reduce): move support types under reduce.ts
