+++ src/operator/min.ts
@@ -1,5 +1,5 @@
 import {Observable} from '../Observable';
-import {ReduceOperator} from './reduce-support';
+import {ReduceOperator} from './reduce';
 
 export function min<T>(comparer?: (value1: T, value2: T) => T): Observable<T> {
   const min: typeof comparer = (typeof comparer === 'function')
